drm/i915: Stage new modeset state straight into atomic state
The logic that stages the state before the modeset was still updating
first the old staged config and then populating the atomic state based
on that. Change this to use only the atomic state.

Note that noew the staged config is updated in the function
intel_modeset_commit_output_state(). This is done so that the modeset
check and hw state force restore continue to work.