drm/i915/tests: Add a test for link training
This adds a test that compiles the link training code from i915 into a
separate executable and uses it to train a fake sink device. The test
iterates over possible combinations of max voltage swing, pre emphasis
and bit rates, with the sink attemting to get the highest values
possible. The test doesn't cover HBR2, however.

Note that the -z muldef option is needed for linking since the test
overrides the implementation of drm_dp_dpcd_write() from the drm dp
helper.