XXX drm/i915: Fix recursive header dependency from i915_trace.h
Since the addition of pipe update trace points in

commit 25ef284a2a7de5393b2ef608046e0d48db0e6e2a
Author: Ville Syrjälä <ville.syrjala@linux.intel.com>
Date:   Tue Apr 29 13:35:48 2014 +0300

    drm/i915: Add pipe update trace points

intel_drv.h includes itself, since it includes i915_drv.h, which
includes i915_trace.h, which in turn includes intel_drv.h.

TODO: figure out if it is possible to break this without creating a new
trace subsystem name.

TODO: check if the move of the TRACE_SYSTEM and TRACE_INCLUDE_FILE to
the unprotected area is really necessary.