drm/i915: Fix Clock recovery sequence
Since

commit cdb0e95bf571dccc1f75fef9bdad21b167ef0b37
Author: Keith Packard <keithp@keithp.com>
Date:   Tue Nov 1 20:00:06 2011 -0700

    drm/i915: Try harder during dp pattern 1 link training

the link training code retries the clock recovery step up to five times
upon failure. This is not part of the DP spec (as noted on that commit's
message) but done to replicate the behavior of "the BIOS". There's no
evidence this improves hardware support, so just implement what the spec
says.

Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Signed-off-by: Ander Conselvan de Oliveira <ander.conselvan.de.oliveira@intel.com>