drm/i915: Limit the number of channel equalizations attempts to 5
In

commit 37f809755845cc3e18e8216c04525bdb885fa13b
Author: Jesse Barnes <jbarnes@virtuousgeek.org>
Date:   Wed Jan 5 14:45:24 2011 -0800

    drm/i915: make DP training try a little harder

attempts to achieve clock recovery were added to the channel
equalization phase of link training. Inspite the claim on that commit's
message, this change doesn't follow the DP spec. The spec requires that
if clock recovery is lost during channel equalization, the whole link
training process be restarted with a lower bit rate. It also says that
up to 5 attempts be done to achieve channel equalization.

This patch removes the extra clock recovery attemps (done after 6 tries of
channel equalization failed) and reduces the number of tries of to 5.
Note that it does not remove the clock recovery step in case CR_DONE is
cleared, and it changes that to count againt the total try count. This
last part is still non-compliant to the DP spec.

Cc: Jesse Barnes <jbarnes@virtuousgeek.org>
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Signed-off-by: Ander Conselvan de Oliveira <ander.conselvan.de.oliveira@intel.com>