drm/i915: Fix link training usage of max vswing and pre-emphasis
From Section 3.5.1.2 Link Training,

When the combination of the requested pre-emphasis level and voltage
swing exceeds the capability of a DisplayPort transmitter, the
transmitter must set the pre-emphasis level according to the request and
use the highest voltage swing it can output with the given pre-emphasis level.

This patch reads the preemphasis requested by panel first during link
training and then updates the voltage swing based on the value supported
as per the pre-emphasis.

Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Signed-off-by: Ander Conselvan de Oliveira <ander.conselvan.de.oliveira@intel.com>