drm/i915: Fix EQ sequence in Link  training
Second half of link training for DP panel is check for equalization.
it is expected that CR should always pass when checking for EQ,
hence this patch fixes this sequence by exiting from function
where EQ is checked if CR fails and returns failure to indicate
the caller of the EQ status.

Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Signed-off-by: Ander Conselvan de Oliveira <ander.conselvan.de.oliveira@intel.com>